rebase abstraction ontop of feed fix changes
