+++ includes/Feed/FeedConfigurationDetectorFactory.php
@@ -1,0 +1,2 @@
+//Todo: write a factory class that initializes the feedconfigurationdetectors, use this file in
+WC_Facebookcommerce and replace the logic there
